requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
azure-common==1.1.28
azure-core==1.29.6
azure-identity==1.12.0
azure-nspkg==3.0.2
azure-search==1.0.0b2
azure-search-documents==11.4.0
azure-storage-blob==12.19.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-spinner==0.1.10
comm==0.2.2
croniter==1.4.1
cryptography==42.0.8
debugpy==1.8.2
decorator==5.1.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastapi-utilities==0.2.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
gotrue==2.5.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.25.2
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
langchain-core==0.2.18
langchain-openai==0.1.16
langsmith==0.1.85
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
msal==1.29.0
msal-extensions==1.2.0
msrest==0.7.1
multidict==6.0.5
nest-asyncio==1.6.0
numpy==2.0.0
oauthlib==3.2.2
openai==1.35.13
orjson==3.10.6
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
portalocker==2.10.0
postgrest==0.13.2
prompt_toolkit==3.0.47
protobuf==5.27.2
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pycparser==2.22
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.8.0
pypdf==3.17.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
realtime==1.0.6
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.19.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.31
stack-data==0.6.3
starlette==0.37.2
storage3==0.7.6
streamlit==1.36.0
StrEnum==0.4.15
supabase==2.3.3
supafunc==0.3.3
tenacity==8.5.0
tiktoken==0.7.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
yarl==1.9.4
